Remove todos and rename
