+++ src/app/exercise/exercises/NotesWithChords/NotesWithChordsExercise.ts
@@ -26,7 +26,7 @@ import {
   IncludedAnswersSettings,
 } from '../utility/settings/IncludedAnswersSettings';
 import { SettingsDescriptors } from '../utility/settings/SettingsDescriptors';
-import { romanNumeralToChordInC } from '../utility/exerciseFactories/BaseRomanAnalysisChordProgressionExercise';
+import { romanNumeralToChordInC } from '../utility/exerciseFactories/romanAnalysisChordProgressionExercise';
 import { scaleDegreeToNoteType } from '../../utility/music/scale-degrees/scaleDegreeToNoteType';
 
 type ChordDegree = 1 | 2 | 3 | 4 | 5 | 6 | 7;
