+++ src/app/exercise/exercises/utility/exerciseFactories/chordProgressionExercise.ts
@@ -10,8 +10,7 @@ import { transpose } from '../../../utility/music/transpose';
 import { Interval } from '../../../utility/music/intervals/Interval';
 import { CreateExerciseParams } from './createExercise';
 
-// todo: drop the "base" and "tonal"
-export type BaseTonalChordProgressionExerciseSettings<GAnswer extends string> = TonalExerciseSettings & {
+export type ChordProgressionExerciseSettings<GAnswer extends string> = TonalExerciseSettings & {
   voiceLeading: 'RANDOM' | 'CORRECT';
   includedPositions: (0 | 1 | 2)[];
   includeBass: boolean;
@@ -31,21 +30,16 @@ export interface ChordProgressionQuestion<GAnswer extends string> extends Omit<E
   }]>;
 }
 
-// todo: remove the class and create those functions instead
-// let's create this method so it doesn't HAVE to be "tonal"
-
 export type ChordProgressionExerciseParams<GAnswer extends string, GSettings extends Exercise.Settings> = {
   getChordProgression: (settings: GSettings) => ChordProgressionQuestion<GAnswer>;
 }
 
-export function chordProgressionExercise<GAnswer extends string, GSettings extends Exercise.Settings>(params: ChordProgressionExerciseParams<GAnswer, GSettings>): CreateExerciseParams<GAnswer, GSettings & BaseTonalChordProgressionExerciseSettings<GAnswer>> {
+export function chordProgressionExercise<GAnswer extends string, GSettings extends Exercise.Settings>(params: ChordProgressionExerciseParams<GAnswer, GSettings>): CreateExerciseParams<GAnswer, GSettings & ChordProgressionExerciseSettings<GAnswer>> {
   throw new Error(`Not implemented`);
 }
 
-// todo: this function will compose chordProgressionExercise and tonalExercise
-// export function tonalChordProgression
-
-export abstract class BaseTonalChordProgressionExercise<GAnswer extends string, GSettings extends BaseTonalChordProgressionExerciseSettings<GAnswer>> extends BaseTonalExercise<GAnswer, GSettings> {
+// todo: remove
+export abstract class BaseTonalChordProgressionExercise<GAnswer extends string, GSettings extends ChordProgressionExerciseSettings<GAnswer>> extends BaseTonalExercise<GAnswer, GSettings> {
   protected override _settings: GSettings = {
     ...this._settings,
     voiceLeading: 'CORRECT',
